pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trail"
version = "0.1.0"
description = "AI-Enabled Progress Tracker + Notion AI Agent + Smart Work Planner"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "G muma", email = "2307556@students.kcau.ac.ke"}
]
dependencies = [
    "psycopg2-binary==2.9.9",
    "redis==5.0.1",
    "python-dotenv==1.0.0",
    "click==8.1.7",
    "PyGithub==2.1.1",
    "notion-client==2.2.1",
    "SQLAlchemy==2.0.23",
    "alembic==1.13.0",
    "rich==13.7.0",
    "python-dateutil==2.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]

[project.scripts]
trail = "src.cli.main:cli"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
